refactor(telemetry): extract shared ingest transport; drop history comment
